Migrate StableSort
